[chore] add comment component 분리
